refactor: readability changes
