+++ apps/api/src/app/shared/framework/swagger/sdk.decorators.ts
@@ -10,6 +10,34 @@ export function SdkGroupName(methodName: string) {
   return applyDecorators(ApiExtension('x-speakeasy-group', methodName));
 }
 // eslint-disable-next-line @typescript-eslint/naming-convention
+export function SdkIgnorePath(methodName: string) {
+  return applyDecorators(ApiExtension('x-speakeasy-ignore', 'true'));
+}
+// eslint-disable-next-line @typescript-eslint/naming-convention
 export function SdkUsageExample(title?: string, description?: string, position?: number) {
   return applyDecorators(ApiExtension('x-speakeasy-usage-example', { title, description, position }));
+}
+
+//eslint-disable-next-line @typescript-eslint/naming-convention
+export function SdkUsePagination(override?: string) {
+  return applyDecorators(
+    ApiExtension('x-speakeasy-pagination', {
+      type: 'offsetLimit',
+      inputs: [
+        {
+          name: 'page',
+          in: 'parameters',
+          type: 'page',
+        },
+        {
+          name: override || 'limit',
+          in: 'parameters',
+          type: 'limit',
+        },
+      ],
+      outputs: {
+        results: '$.data.resultArray',
+      },
+    })
+  );
 }
